+++ test/semanticValidatorTests.ts
@@ -71,7 +71,8 @@ describe("Semantic validation", () => {
     it("should pass when validating a swagger with using arm-id format for string type", async () => {
       const specPath = `${testPath}/semanticValidation/specification/validateCompile/Swagger-with-xms-extension.json`;
       const result = await validate.validateSpec(specPath, undefined);
-      assert(result.validityStatus === true,
+      assert(
+        result.validityStatus === true,
         `swagger "${specPath}" contains semantic validation errors.`
       );
     });
@@ -285,6 +286,14 @@ describe("Semantic validation", () => {
       const specPath = `${testPath}/semanticValidation/specification/discriminator/notRequiredDiscriminatorWithSuppression.json`;
       const result = await validate.validateSpec(specPath, undefined);
       assert(result.validityStatus === true);
+    });
+    it("should validate without additionalProperty Error on the refWithReadOnly property injected by oav", async () => {
+      const specPath = `${testPath}/modelValidation/swaggers/specification/refWithReadOnlyProperyError/cognitiveservices.json`;
+      const result = await validate.validateSpec(specPath, undefined);
+      assert(
+        result.validityStatus === true,
+        `swagger "${specPath}" contains semantic validation errors.`
+      );
     });
   });
 });